practica

+++ 2do_cuatrimestre/ejercicios_practica/clase_16/persona.h
@@ -2,7 +2,7 @@ typedef struct{
 
     int edad;
     char nombre[50];
-    int dni;
+    long int dni;
 
 }S_Person;
